\docType{data}
\name{RegionB}
\alias{RegionB}
\title{Haplotype counts around gene HSA2 in domestic dogs}
\format{A data.frame with 30 rows and 124 columns. Each row is a dog breed and each column is an haplotype.}
\source{
Friguet, C. and Emily, M. Sélection de marqueurs pour la détection d'interactions de gènes. Proceedings of 45èmes Journées de Statistique, May 2013, Toulouse, France. pp.178, 2013.\cr
}
\usage{
data(RegionA)
}
\description{
	A matrix-like dataset containing the simulation of an haplotype of a 5 five locus long genomic region. A total of 1000 individuals have been simulated. Each locus of the haplotype has two possible outcomes, coded as 0 or 1.
}
\value{
A dataset containing 1000 five locus long haplotype.}

\keyword{datasets}

\seealso{
\code{\link{RegionA}}
}

